chore(dev.yml): update Java setup action to use Temurin distribution and specify Java version as '17' chore(ops.yml): comment out the 'on: [push]' trigger to disable the workflow chore(sec.yml): comment out the 'on: [push]' trigger to disable the workflow
The Java setup action in the dev.yml workflow has been updated to use the Temurin distribution and specify the Java version as '17'. This ensures that the correct Java version is used for the build process. The 'on: [push]' triggers in the ops.yml and sec.yml workflows have been commented out to disable the workflows for now.