feat: add repository configuration for Maven GitHub Package Registry
A new file `repository.gradle.kts` is added to configure the repositories for Gradle. This file sets up the necessary repositories for accessing dependencies from the Maven GitHub Package Registry. The `gradlePluginPortal()` and the GitHub Package Registry repository are added to the `pluginManagement` block, while the GitHub Package Registry repository is added to the `allprojects` block.

In the workflow file `dev-workflow.yml`, a new step is added to copy the `repository.gradle.kts` file to the Gradle initialization directory before running the setup for Maven GitHub Package Registry. This ensures that the repository configuration is applied correctly during the build process.

These changes enable the project to resolve dependencies from the Maven GitHub Package Registry, allowing for easier management and distribution of packages within the project.