chore(jvm): update action.yml to add support for custom Java distribution and fix variable naming inconsistency
The action.yml file has been updated to include a new input `java-distribution` which allows users to specify a custom Java distribution. This input is optional and has a default value of 'temurin'. Additionally, the variable naming inconsistency has been fixed by changing `java-Version` to `java-Version` in the `Setup Java` step and `java-version` to `java-Version` in the `Setup Java` and `Setup Gradle` steps. This improves the clarity and consistency of the code.