chore(chromatic): update action.yml and workflows to support project token input
The action.yml file has been updated to include a new input called "project-token" which is used to specify the path to the artifact to upload. This input is now required and can be left empty if no artifact is needed.

The make-jvm-workflow.yml and publish-chromatic-workflow.yml files have been updated to include the "CHROMATIC_PROJECT_TOKEN" input, which is now required and can be left empty. This input is used in the chromatic action to specify the project token.

The publish-storybook-workflow.yml file has also been updated to include the "CHROMATIC_PROJECT_TOKEN" input, which is now required and can be left empty. This input is used in the package-storybook job to specify the project token.