fix(action.yml): update action.yml to use correct action path for make-step-optional action feat(action.yml): add new action.yml for version action to retrieve project version from VERSION file fix(make-jvm-workflow.yml): remove unnecessary environment variables and add step to get version from file fix(make-nodejs-workflow.yml): remove unnecessary environment variables and add step to get version from file
The changes in the action.yml files update the uses path for the make-step-optional action to the correct location. This ensures that the action is correctly referenced and executed.

The addition of the version action.yml introduces a new action that retrieves the project version from the VERSION file. This allows for the version information to be displayed in the workflow summary.

In the make-jvm-workflow.yml file, unnecessary environment variables are removed, and a new step is added to get the version from the file.

Similarly, in the make-nodejs-workflow.yml file, unnecessary environment variables are removed, and a new step is added to get the version from the file.

chore(publish-storybook-workflow.yml): remove unnecessary 'with-package-task' input
The 'with-package-task' input is no longer needed and has been removed from the workflow configuration. This change simplifies the configuration and improves clarity.