feat: add setup for npm with GitHub Packages in GitHub Actions workflows
The new action `setup-npm-github-pkg` is added to configure npm to use a GitHub Packages repository for dependencies. This allows for better management of dependencies and ensures that the correct packages are used in the workflow. The action generates a `.npmrc` file with the necessary configurations for npm to work with the GitHub Packages repository. Additionally, the workflow file `make-nodejs-workflow.yml` is updated to include the setup for npm with GitHub Packages by conditionally running the `setup-npm-github-pkg` action based on the input flag `with-setup-npm-github-pkg`. The action is provided with the npm authentication token from the secrets. The workflow also includes a step for Docker registry login if the flag `with-docker-registry-login` is set to true.

The `publish-storybook-workflow.yml` file is updated to set the `with-setup-npm-github-pkg` flag to false, indicating that the npm setup with GitHub Packages is not required for this specific workflow. The npm authentication token is passed as `NPM_PKG_GITHUB_TOKEN` from the secrets.