feat(version): update action.yml to manage version tags for different branches and snapshots feat(workflow): update make-nodejs-workflow.yml to use '-alpha-SNAPSHOT' as pre-release tag for the main branch
The changes in action.yml allow for managing version tags for different branches and snapshots by introducing new input options for pre-release tags for main, release, feature, fix, and refactoring branches. This provides more flexibility in versioning strategies based on the branch type.

The update in make-nodejs-workflow.yml sets the pre-release tag '-alpha-SNAPSHOT' for the main branch, aligning the workflow with the versioning strategy defined in the action.yml file.