feat(setup-npm-github-pkg): add support for custom npm registry in setup action
feat(make-nodejs-workflow): enable setup for npm with GitHub Packages and npmjs

The setup-npm-github-pkg action now includes a new input `npm-registry` to allow users to specify a custom npm registry. This change provides flexibility for users who need to use a different npm registry other than the default one.

In the make-nodejs-workflow, a new step has been added to set up npm with GitHub Packages and npmjs based on the branch conditions. This enhancement ensures that the workflow can handle publishing to both GitHub Packages and npmjs registry based on the branch being worked on.