feat(make-jvm-workflow.yml): add support for Docker login to the Docker registry
The workflow file now includes a new input parameter 'with-docker-registry-login' which is a flag to enable Docker login to the Docker registry. When this flag is set to 'true', a step for Docker registry login is executed using the docker/login-action@v3 GitHub Action. This enhancement allows for seamless integration with Docker registries for the workflow when needed.