ci(workflows): add conditional checks to GitHub Actions workflows
Add conditional checks to ensure certain tasks only run under specific
conditions. This prevents unnecessary execution of tasks and optimizes
workflow runs. For example, the `run-dev-tasks` job now only runs if the
branch name does not start with 'versionning/' or if the event is not a
pull request. Similarly, make tasks like lint, test, and check now have
conditions to run only when appropriate, such as on specific tags or
branches.