chore(dependencies): update various dependencies to their latest versions
Update dependencies to ensure the project uses the latest versions,
which include bug fixes, performance improvements, and new features.
Specifically, update GraalVM to 0.10.3, SLF4J to 2.0.16, Jakarta Persistence
to 3.2.0, Reactor to 3.6.10, Jackson to 2.18.0, Cucumber to 7.19.0, JUnit
to 5.11.1, JUnit Platform to 1.11.1, Testcontainers to 1.20.2, and Kotlin
Datetime to 0.6.1. This helps maintain the project's stability and security.